Move the routing API from Hedging to root folder
