+++ src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/Endpoint.cs
