+++ src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/IRequestRoutingStrategy.cs
