+++ src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/IRoutingStrategyBuilder.cs
