+++ src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/Internal/IPooledRequestRoutingStrategyFactory.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.Extensions.Http.Resilience.Internal.Routing;
+namespace Microsoft.Extensions.Http.Resilience.Routing.Internal;
 
 /// <inheritdoc/>
 internal interface IPooledRequestRoutingStrategyFactory : IRequestRoutingStrategyFactory
