+++ src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/Internal/RoutingStrategyBuilder.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.Extensions.Http.Resilience.Internal.Routing;
+namespace Microsoft.Extensions.Http.Resilience.Routing.Internal;
 
 internal sealed class RoutingStrategyBuilder : IRoutingStrategyBuilder
 {
