+++ src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/WeightedEndpoint.cs
