+++ test/Libraries/Microsoft.Extensions.Http.Resilience.Tests/Routing/WeightedRoutingStrategyTest.cs
@@ -6,11 +6,11 @@ using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
 using Microsoft.Extensions.Http.Resilience.Internal;
-using Microsoft.Extensions.Http.Resilience.Internal.Routing;
+using Microsoft.Extensions.Http.Resilience.Routing.Internal.WeightedGroups;
 using Moq;
 using Xunit;
 
-namespace Microsoft.Extensions.Http.Resilience.Test.Hedging.Routing;
+namespace Microsoft.Extensions.Http.Resilience.Test.Routing;
 
 public class WeightedRoutingStrategyTest : RoutingStrategyTest
 {
